# src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mithra Direct App",
    "description": "API documentation for the subscription services",
    "contact": {
      "name": "Swamy Kunta",
      "url": "https://github.com/s713278/ECommerceApp/tree/master",
      "email": "swamy.kunta@gmail.com"
    },
    "license": {
      "name": "License",
      "url": "/"
    },
    "version": "v1"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Env"
    },
    {
      "url": "http://172.105.33.153:8080/api",
      "description": "Dev Env"
    }
  ],
  "paths": {
...
    "/v1/users/{userId}/subs/": {
      "post": {
        "tags": [
          "3. Subscription Management"
        ],
        "summary": "Create Subscription",
        "operationId": "createSubscription",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
    "/v1/users/{userId}/subs/{subId}": {
      "get": {
        "tags": [
          "3. Subscription Management"
        ],
        "summary": "Fetch Subscription Details",
        "operationId": "fetchSubscription",
        "parameters": [
          {
            "name": "subId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      },
      "delete": {
        "tags": [
          "3. Subscription Management"
        ],
        "summary": "Delete Subscription",
        "operationId": "deleteSubscription",
        "parameters": [
          {
            "name": "subId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      },
      "patch": {
        "tags": [
          "3. Subscription Management"
        ],
        "summary": "Update Subscription",
        "operationId": "updateSubscription",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "subId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
    "/v1/users/{userId}/subs/{subId}/status": {
      "patch": {
        "tags": [
          "3. Subscription Management"
        ],
        "description": "Update subscription status",
        "operationId": "updateSubscriptionStatus",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "subId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
    "/v1/vendors/{vendorId}/subs": {
      "get": {
        "tags": [
          "2. Vendors/Product Listing"
        ],
        "summary": "Fetch Subscriptions By Vendor ID",
        "operationId": "fetchSubsByVendorId",
        "parameters": [
          {
            "name": "vendorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/users/{userId}/subs": {
      "get": {
        "tags": [
          "5. User Management"
        ],
        "summary": "Fetch Subscriptions By User ID",
        "operationId": "fetchAllSubsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponseObject": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Customer registered successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "description": "HTTP Status Code",
            "format": "int32",
            "example": 201
          },
          "data": {
            "type": "object",
            "description": "Data"
          }
        }
      },
...
      "CreateSubscriptionRequest": {
        "required": [
          "frequency",
          "price_id",
          "quantity",
          "sku_id"
        ],
        "type": "object",
        "properties": {
          "quantity": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "ONE_TIME",
              "DAILY",
              "ALTERNATE_DAY",
              "WEEKLY",
              "CUSTOM"
            ]
          },
          "sku_id": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          },
          "price_id": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          },
          "delivery_date": {
            "type": "string",
            "format": "date"
          },
          "start_date": {
            "type": "string",
            "format": "date"
          },
          "end_date": {
            "type": "string",
            "format": "date"
          },
          "custom_days": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "Data": {
        "type": "object",
        "properties": {
          "subscription_id": {
            "type": "integer",
            "format": "int64"
          },
          "next_delivery_date": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "SubscriptionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Data"
          }
        }
      },
...
      "UpdateSubscriptionRequest": {
        "required": [
          "frequency",
          "quantity",
          "sku_id"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACTIVE",
              "PAUSED",
              "CANCELLED",
              "EXPIRED",
              "DELETED"
            ]
          },
          "quantity": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "ONE_TIME",
              "DAILY",
              "ALTERNATE_DAY",
              "WEEKLY",
              "CUSTOM"
            ]
          },
          "sku_id": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          },
          "delivery_date": {
            "type": "string",
            "format": "date"
          },
          "start_date": {
            "type": "string",
            "format": "date"
          },
          "end_date": {
            "type": "string",
            "format": "date"
          },
          "custom_days": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "SubscriptionStatusRequest": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status of the subscription",
            "example": "NEW",
            "enum": [
              "PENDING",
              "ACTIVE",
              "PAUSED",
              "CANCELLED",
              "EXPIRED",
              "DELETED"
            ]
          }
        }
      },
...
      "Subscription": {
        "type": "object",
        "properties": {
          "createdBy": {
            "type": "string",
            "writeOnly": true
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "skuId": {
            "type": "integer",
            "format": "int64"
          },
          "priceId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "ONE_TIME",
              "DAILY",
              "ALTERNATE_DAY",
              "WEEKLY",
              "CUSTOM"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACTIVE",
              "PAUSED",
              "CANCELLED",
              "EXPIRED",
              "DELETED"
            ]
          },
          "customDays": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "nextDeliveryDate": {
            "type": "string",
            "format": "date"
          },
          "deliveryAddress": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "updateVersion": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "APIResponseString": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Customer registered successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "description": "HTTP Status Code",
            "format": "int32",
            "example": 201
          },
          "data": {
            "type": "string",
            "description": "Data"
          }
        }
      }
    },
    "securitySchemes": {
      "E-Commerce Application": {
        "type": "http",
        "in": "header",
        "scheme": "bearer"
      }
    }
  }
}